client/requirements.txt
kafka-python==2.0.2
locust==2.5.1
numpy==1.21.4
google-cloud-monitoring==2.9.1
pandas==1.3.4
pyzmq==22.3.0
